// src/styles/types.ts
export enum Alignment {
  center = 'center',
  left = 'left',
  right = 'right',
}

export enum Size {
  small = 'small',
  medium = 'medium',
  large = 'large',
  extralarge = 'extralarge',
}

export enum Variant {
  primary = 'primary',
  secondary = 'secondary',
  desctructionary = 'desctructionary',
  text = 'text',
}

export enum Position {
  left = 'left',
  right = 'right',
}